refactor: reorganize code to remove utils
