+++ packages/eip712/src/payload/common.ts
@@ -2,11 +2,6 @@ export interface JSONObject {
   [key: string]: any
 }
 
-export interface EIP712Type {
-  type: string
-  name: string
-}
-
 export interface MessageParams {
   payload: JSONObject
   numMessages: number
